Fixed compiler settings
I had created a command to set the compiler for tps, but it was not
working that well. It was buffer local and had to be explicitly set.

Modified it so its set for every .cpp/h buffer we load. Removed the
command and just set the compiler directly in the autocmd.

The tps compiler plugin also had some logging. This causes the compiler
to pause on the messages, requiring a <cr> to continue. Removed the
logging.

Finally, we had some functions and mapping for support which was not
very helpful. Removed those, and instead added some mappings for
navigating the quickfix list.